refactor(ui): 重构 SVG 图标组件
- 将组件名称从 SvgIcon 改为 KSvgIcon
- 为组件添加 defineOptions，设置组件名称
- 优化类型定义，使用 interface 替代 type
- 调整颜色配置，使用硬编码的值替代 CSS 变量
- 优化模板中的条件判断，使用可选链操作符
