+++ packages/ui/src/components/k-svg-icon/index.ts
@@ -1,5 +1,5 @@
 import { App } from "vue"
-import SvgIcon from "./svg-icon.vue"
+import SvgIcon from "./k-svg-icon.vue"
 
 SvgIcon.install = (app: App) => {
   app.component(SvgIcon.name!, SvgIcon)
